test/test_zip2tar: Make tar time test work for different timezones.
Unfortunately, the dates output by tar are timezone dependent
and they may change accordingly. To cope with that, we use GNU tar's
`--utc` option (not really sure if other tar implementations support
such options).

While at it, I also fixed the strings to really be in UTC.

This makes the test pass on my timezone (-0300).

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>